fix: corrige tipo do price no model e migration

+++ back-end/src/database/models/Product.js
@@ -2,7 +2,7 @@ const Products = (sequelize, DataTypes) => {
   const Product = sequelize.define('Product', {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     name: { type: DataTypes.STRING, unique: true },
-    price: DataTypes.INTEGER,
+    price: DataTypes.DECIMAL,
     urlImage: DataTypes.STRING,
   },
     {
